Adjusting to conform to android signature to avoid method not found errors

+++ compat/android/database/AbstractCursor.java
@@ -113,8 +113,8 @@ public class AbstractCursor implements Cursor {
     return SQLite.sqlite3_column_blob(pStmt, ci).getByteArray(0,  SQLite.sqlite3_column_bytes(pStmt, ci));
   }
 
-  public byte getShort(int ci) {
-    return (byte)getInt(ci);
+  public short getShort(int ci) {
+    return (short)getInt(ci);
   }
 
   public double getDouble(int ci) {
